Added collisions

+++ src/player/player_rotate.c
@@ -6,7 +6,7 @@
 /*   By: bbrassar <bbrassar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/26 13:25:54 by bbrassar          #+#    #+#             */
-/*   Updated: 2022/06/02 09:08:28 by bbrassar         ###   ########.fr       */
+/*   Updated: 2022/06/02 14:56:54 by bbrassar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 #include <math.h>
 
-void	player_rotate(t_cub *cub, int yaw)
+void	player_rotate(t_cub *cub, float yaw)
 {
 	if (yaw != 0)
 	{
